+++ bootlaces/src/main/java/com/candroid/bootlaces/BootService.kt
@@ -14,17 +14,8 @@ limitations under the License.*/
 package com.candroid.bootlaces
 
 import android.app.Service
-import android.content.BroadcastReceiver
-import android.content.Context
 import android.content.Intent
-import android.content.IntentFilter
-import android.os.Build
 import android.os.IBinder
-import androidx.lifecycle.LifecycleService
-import androidx.lifecycle.ServiceLifecycleDispatcher
-import androidx.lifecycle.lifecycleScope
-import androidx.localbroadcastmanager.content.LocalBroadcastManager
-import kotlinx.coroutines.runBlocking
 
 /*
             (   (                ) (             (     (
@@ -50,6 +41,8 @@ import kotlinx.coroutines.runBlocking
  * @author Chris Basinger
  * @email evilthreads669966@gmail.com
  * @date 10/09/20
+ *
+ * Persistent foreground service that is started by the startBoot method and BootReceiver
  **/
 abstract class BootService : Service() {
     private lateinit var notifProxy: NotificationProxy
@@ -76,102 +69,3 @@ abstract class BootService : Service() {
 }
 
 
-abstract class LifecycleBootService: LifecycleService() {
-    private val mDispatcher = ServiceLifecycleDispatcher(this)
-    private lateinit var notifProxy: NotificationProxy
-
-    @PublishedApi
-    internal companion object{
-        var payload: (suspend () -> Unit)? = null
-    }
-
-    init {
-        if(payload != null)
-            lifecycleScope.launchWhenCreated {
-                payload!!.invoke()
-            }
-    }
-
-    override fun getLifecycle() = mDispatcher.lifecycle
-
-    override fun onCreate() {
-        mDispatcher.onServicePreSuperOnCreate()
-        BootServiceState.setRunning()
-        super.onCreate()
-        notifProxy = NotificationProxy()
-        notifProxy.onCreate(this)
-    }
-
-    override fun onStart(intent: Intent?, startId: Int) {
-        mDispatcher.onServicePreSuperOnStart()
-        super.onStart(intent, startId)
-    }
-
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        super.onStartCommand(intent, flags, startId)
-        return START_STICKY
-    }
-
-    override fun onDestroy() {
-        mDispatcher.onServicePreSuperOnDestroy()
-        notifProxy.onDestroy(this)
-        super.onDestroy()
-        BootServiceState.setStopped()
-    }
-}
-
-internal class NotificationProxy{
-    private lateinit var receiver: UpdateReceiver
-
-    fun onCreate(ctx: Service){
-        if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O){
-            val filter = IntentFilter(Actions.ACTION_UPDATE)
-            receiver = UpdateReceiver()
-            LocalBroadcastManager.getInstance(ctx).registerReceiver(receiver, filter)
-            runBlocking {
-                startBootNotification(ctx)
-            }
-        }
-    }
-
-    fun onDestroy(ctx: Service){
-        if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
-            LocalBroadcastManager.getInstance(ctx).unregisterReceiver(receiver)
-    }
-
-    /*create boot service notification*/
-    private suspend fun startBootNotification(ctx: Service): Boolean {
-        BootNotificationFactory.Configuration.createChannel(ctx)
-        val notification = BootNotificationFactory.getInstance(ctx).createNotification()
-        notification?.let { notif ->
-            ctx.startForeground(BootNotificationFactory.Configuration.FOREGROUND_ID, notif)
-            return true
-        }
-        return false
-    }
-
-    inner class UpdateReceiver: BroadcastReceiver(){
-
-        override fun onReceive(ctx: Context?, intent: Intent?) {
-            if(intent?.action?.equals(Actions.ACTION_UPDATE) ?: false){
-                val boot = Boot()
-
-                if(intent!!.hasExtra(BootRepository.KEY_ACTIVITY))
-                    boot.activity = intent.getStringExtra(BootRepository.KEY_ACTIVITY)
-
-                if(intent.hasExtra(BootRepository.KEY_TITLE))
-                    boot.title = intent.getStringExtra(BootRepository.KEY_TITLE)
-
-                if(intent.hasExtra(BootRepository.KEY_CONTENT))
-                    boot.content = intent.getStringExtra(BootRepository.KEY_CONTENT)
-
-                if(intent.hasExtra(BootRepository.KEY_ICON))
-                    boot.icon = intent.getIntExtra(BootRepository.KEY_ICON, -1).takeIf { ic -> ic != -1 }
-
-                runBlocking {
-                    BootNotificationFactory.getInstance(ctx!!).updateBootNotification(boot)
-                }
-            }
-        }
-    }
-}
